fixed calloc, bzero

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: fzaazaa <fzaazaa@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/10 04:54:29 by fzaazaa           #+#    #+#             */
-/*   Updated: 2024/11/10 05:20:15 by fzaazaa          ###   ########.fr       */
+/*   Updated: 2024/11/23 14:20:54 by fzaazaa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// voir issue github
+#include"libft.h"
 
-void	ft_bzero(void *s, unsigned int size)
+void	ft_bzero(void *s, size_t size)
 {
 	unsigned int	i;
